separated iex musedata module
