--- src/importexport/internal/musedata/musedata.cmake
@@ -1,5 +1,0 @@
-
-set (MUSEDATA_SRC
-    ${CMAKE_CURRENT_LIST_DIR}/musedata.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/musedata.h
-)
